+++ src/zarr/v3/codecs/bytes.py
@@ -13,15 +13,17 @@ from attr import frozen, field
 from zarr.v3.abc.codec import ArrayBytesCodec
 from zarr.v3.codecs.registry import register_codec
 from zarr.v3.common import BytesLike
-from zarr.v3.metadata import CodecMetadata
 
 if TYPE_CHECKING:
-    from zarr.v3.metadata import CoreArrayMetadata
+    from zarr.v3.metadata import CodecMetadata, ArraySpec, ArrayMetadata, RuntimeConfiguration
+
+
+Endian = Literal["big", "little"]
 
 
 @frozen
 class BytesCodecConfigurationMetadata:
-    endian: Optional[Literal["big", "little"]] = "little"
+    endian: Optional[Endian] = "little"
 
 
 @frozen
@@ -32,28 +34,24 @@ class BytesCodecMetadata:
 
 @frozen
 class BytesCodec(ArrayBytesCodec):
-    array_metadata: CoreArrayMetadata
     configuration: BytesCodecConfigurationMetadata
     is_fixed_size = True
 
     @classmethod
-    def from_metadata(
-        cls, codec_metadata: CodecMetadata, array_metadata: CoreArrayMetadata
-    ) -> BytesCodec:
+    def from_metadata(cls, codec_metadata: CodecMetadata) -> BytesCodec:
         assert isinstance(codec_metadata, BytesCodecMetadata)
-        assert (
-            array_metadata.dtype.itemsize == 1 or codec_metadata.configuration.endian is not None
-        ), "The `endian` configuration needs to be specified for multi-byte data types."
-        return cls(
-            array_metadata=array_metadata,
-            configuration=codec_metadata.configuration,
-        )
+        return cls(configuration=codec_metadata.configuration)
 
     @classmethod
     def get_metadata_class(cls) -> Type[BytesCodecMetadata]:
         return BytesCodecMetadata
 
-    def _get_byteorder(self, array: np.ndarray) -> Literal["big", "little"]:
+    def validate(self, array_metadata: ArrayMetadata) -> None:
+        assert (
+            not array_metadata.data_type.has_endianness or self.configuration.endian is not None
+        ), "The `endian` configuration needs to be specified for multi-byte data types."
+
+    def _get_byteorder(self, array: np.ndarray) -> Endian:
         if array.dtype.byteorder == "<":
             return "little"
         elif array.dtype.byteorder == ">":
@@ -66,27 +64,31 @@ class BytesCodec(ArrayBytesCodec):
     async def decode(
         self,
         chunk_bytes: BytesLike,
+        chunk_spec: ArraySpec,
+        _runtime_configuration: RuntimeConfiguration,
     ) -> np.ndarray:
-        if self.array_metadata.dtype.itemsize > 0:
+        if chunk_spec.dtype.itemsize > 0:
             if self.configuration.endian == "little":
                 prefix = "<"
             else:
                 prefix = ">"
-            dtype = np.dtype(f"{prefix}{self.array_metadata.data_type.to_numpy_shortname()}")
+            dtype = np.dtype(f"{prefix}{chunk_spec.data_type.to_numpy_shortname()}")
         else:
-            dtype = np.dtype(f"|{self.array_metadata.data_type.to_numpy_shortname()}")
+            dtype = np.dtype(f"|{chunk_spec.data_type.to_numpy_shortname()}")
         chunk_array = np.frombuffer(chunk_bytes, dtype)
 
         # ensure correct chunk shape
-        if chunk_array.shape != self.array_metadata.chunk_shape:
+        if chunk_array.shape != chunk_spec.shape:
             chunk_array = chunk_array.reshape(
-                self.array_metadata.chunk_shape,
+                chunk_spec.shape,
             )
         return chunk_array
 
     async def encode(
         self,
         chunk_array: np.ndarray,
+        _chunk_spec: ArraySpec,
+        _runtime_configuration: RuntimeConfiguration,
     ) -> Optional[BytesLike]:
         if chunk_array.dtype.itemsize > 1:
             byteorder = self._get_byteorder(chunk_array)
@@ -95,7 +97,7 @@ class BytesCodec(ArrayBytesCodec):
                 chunk_array = chunk_array.astype(new_dtype)
         return chunk_array.tobytes()
 
-    def compute_encoded_size(self, input_byte_length: int) -> int:
+    def compute_encoded_size(self, input_byte_length: int, _chunk_spec: ArraySpec) -> int:
         return input_byte_length
