+++ src/zarr/v3/codecs/gzip.py
@@ -13,10 +13,9 @@ from numcodecs.gzip import GZip
 from zarr.v3.abc.codec import BytesBytesCodec
 from zarr.v3.codecs.registry import register_codec
 from zarr.v3.common import BytesLike, to_thread
-from zarr.v3.metadata import CodecMetadata
 
 if TYPE_CHECKING:
-    from zarr.v3.metadata import CoreArrayMetadata
+    from zarr.v3.metadata import ArraySpec, CodecMetadata, RuntimeConfiguration
 
 
 @frozen
@@ -32,20 +31,14 @@ class GzipCodecMetadata:
 
 @frozen
 class GzipCodec(BytesBytesCodec):
-    array_metadata: CoreArrayMetadata
     configuration: GzipCodecConfigurationMetadata
     is_fixed_size = True
 
     @classmethod
-    def from_metadata(
-        cls, codec_metadata: CodecMetadata, array_metadata: CoreArrayMetadata
-    ) -> GzipCodec:
+    def from_metadata(cls, codec_metadata: CodecMetadata) -> GzipCodec:
         assert isinstance(codec_metadata, GzipCodecMetadata)
 
-        return cls(
-            array_metadata=array_metadata,
-            configuration=codec_metadata.configuration,
-        )
+        return cls(configuration=codec_metadata.configuration)
 
     @classmethod
     def get_metadata_class(cls) -> Type[GzipCodecMetadata]:
@@ -54,16 +47,24 @@ class GzipCodec(BytesBytesCodec):
     async def decode(
         self,
         chunk_bytes: bytes,
+        _chunk_spec: ArraySpec,
+        _runtime_configuration: RuntimeConfiguration,
     ) -> BytesLike:
         return await to_thread(GZip(self.configuration.level).decode, chunk_bytes)
 
     async def encode(
         self,
         chunk_bytes: bytes,
+        _chunk_spec: ArraySpec,
+        _runtime_configuration: RuntimeConfiguration,
     ) -> Optional[BytesLike]:
         return await to_thread(GZip(self.configuration.level).encode, chunk_bytes)
 
-    def compute_encoded_size(self, _input_byte_length: int) -> int:
+    def compute_encoded_size(
+        self,
+        _input_byte_length: int,
+        _chunk_spec: ArraySpec,
+    ) -> int:
         raise NotImplementedError
