+++ src/zarr/v3/codecs/transpose.py
@@ -9,14 +9,13 @@ from typing import (
 )
 
 import numpy as np
-from attr import frozen, field
+from attr import evolve, frozen, field
 
 from zarr.v3.abc.codec import ArrayArrayCodec
 from zarr.v3.codecs.registry import register_codec
-from zarr.v3.metadata import CodecMetadata
 
 if TYPE_CHECKING:
-    from zarr.v3.metadata import CoreArrayMetadata
+    from zarr.v3.metadata import ArraySpec, CodecMetadata, RuntimeConfiguration
 
 
 @frozen
@@ -32,69 +31,60 @@ class TransposeCodecMetadata:
 
 @frozen
 class TransposeCodec(ArrayArrayCodec):
-    array_metadata: CoreArrayMetadata
     order: Tuple[int, ...]
     is_fixed_size = True
 
     @classmethod
-    def from_metadata(
-        cls, codec_metadata: CodecMetadata, array_metadata: CoreArrayMetadata
-    ) -> TransposeCodec:
+    def from_metadata(cls, codec_metadata: CodecMetadata) -> TransposeCodec:
         assert isinstance(codec_metadata, TransposeCodecMetadata)
+        return cls(order=codec_metadata.configuration.order)
 
-        configuration = codec_metadata.configuration
+    def evolve(self, *, ndim: int, **_kwargs) -> TransposeCodec:
         # Compatibility with older version of ZEP1
-        if configuration.order == "F":  # type: ignore
-            order = tuple(array_metadata.ndim - x - 1 for x in range(array_metadata.ndim))
+        if self.order == "F":  # type: ignore
+            order = tuple(ndim - x - 1 for x in range(ndim))
 
-        elif configuration.order == "C":  # type: ignore
-            order = tuple(range(array_metadata.ndim))
+        elif self.order == "C":  # type: ignore
+            order = tuple(range(ndim))
 
         else:
-            assert len(configuration.order) == array_metadata.ndim, (
+            assert len(self.order) == ndim, (
                 "The `order` tuple needs have as many entries as "
-                + f"there are dimensions in the array. Got: {configuration.order}"
+                + f"there are dimensions in the array. Got: {self.order}"
             )
-            assert len(configuration.order) == len(set(configuration.order)), (
-                "There must not be duplicates in the `order` tuple. "
-                + f"Got: {configuration.order}"
+            assert len(self.order) == len(set(self.order)), (
+                "There must not be duplicates in the `order` tuple. " + f"Got: {self.order}"
             )
-            assert all(0 <= x < array_metadata.ndim for x in configuration.order), (
+            assert all(0 <= x < ndim for x in self.order), (
                 "All entries in the `order` tuple must be between 0 and "
-                + f"the number of dimensions in the array. Got: {configuration.order}"
+                + f"the number of dimensions in the array. Got: {self.order}"
             )
-            order = tuple(configuration.order)
+            order = tuple(self.order)
 
-        return cls(
-            array_metadata=array_metadata,
-            order=order,
-        )
+        if order != self.order:
+            return evolve(self, order=order)
+        return self
 
     @classmethod
     def get_metadata_class(cls) -> Type[TransposeCodecMetadata]:
         return TransposeCodecMetadata
 
-    def resolve_metadata(self) -> CoreArrayMetadata:
-        from zarr.v3.metadata import CoreArrayMetadata
+    def resolve_metadata(self, chunk_spec: ArraySpec) -> ArraySpec:
+        from zarr.v3.metadata import ArraySpec
 
-        return CoreArrayMetadata(
-            shape=tuple(
-                self.array_metadata.shape[self.order[i]] for i in range(self.array_metadata.ndim)
-            ),
-            chunk_shape=tuple(
-                self.array_metadata.chunk_shape[self.order[i]]
-                for i in range(self.array_metadata.ndim)
-            ),
-            data_type=self.array_metadata.data_type,
-            fill_value=self.array_metadata.fill_value,
-            runtime_configuration=self.array_metadata.runtime_configuration,
+        return ArraySpec(
+            shape=tuple(chunk_spec.shape[self.order[i]] for i in range(chunk_spec.ndim)),
+            data_type=chunk_spec.data_type,
+            fill_value=chunk_spec.fill_value,
         )
 
     async def decode(
         self,
         chunk_array: np.ndarray,
+        chunk_spec: ArraySpec,
+        _runtime_configuration: RuntimeConfiguration,
     ) -> np.ndarray:
-        inverse_order = [0 for _ in range(self.array_metadata.ndim)]
+        inverse_order = [0] * chunk_spec.ndim
         for x, i in enumerate(self.order):
             inverse_order[x] = i
         chunk_array = chunk_array.transpose(inverse_order)
@@ -103,11 +93,13 @@ class TransposeCodec(ArrayArrayCodec):
     async def encode(
         self,
         chunk_array: np.ndarray,
+        chunk_spec: ArraySpec,
+        _runtime_configuration: RuntimeConfiguration,
     ) -> Optional[np.ndarray]:
         chunk_array = chunk_array.transpose(self.order)
         return chunk_array
 
-    def compute_encoded_size(self, input_byte_length: int) -> int:
+    def compute_encoded_size(self, input_byte_length: int, _chunk_spec: ArraySpec) -> int:
         return input_byte_length
